requirements-security.txt
# REVENG Universal Reverse Engineering Platform - Security-Focused Dependencies
# =============================================================================
# This file contains security-hardened versions of dependencies.
# Use this for production deployments and security-sensitive environments.

# =============================================================================
# CORE DEPENDENCIES - Security-hardened versions
# =============================================================================

# HTTP client for API calls (security updates)
requests>=2.31.0,<3.0.0

# Ghidra MCP integration for professional disassembly
ghidramcp>=0.1.0,<1.0.0

# =============================================================================
# BINARY MANIPULATION - Security-hardened versions
# =============================================================================

# Binary parsing and patching (security updates)
lief>=0.13.2,<0.14.0

# Multi-architecture assembler (security updates)
keystone-engine>=0.9.2,<1.0.0

# Multi-architecture disassembler (security updates)
capstone>=5.0.1,<6.0.0

# =============================================================================
# VISUALIZATION AND ANALYSIS - Security-hardened versions
# =============================================================================

# Control flow graph analysis (security updates)
networkx>=3.1,<4.0.0

# Graph visualization (security updates)
pydot>=1.4.2,<2.0.0

# Progress bars for long-running operations (security updates)
tqdm>=4.66.0,<5.0.0

# Configuration file support (security updates)
pyyaml>=6.0.1,<7.0.0

# =============================================================================
# SECURITY REPLACEMENTS FOR VULNERABLE DEPENDENCIES
# =============================================================================

# Replace pypdf2 with pypdf (actively maintained, secure)
pypdf>=3.17.0,<4.0.0

# Replace ecdsa with cryptography (more secure, actively maintained)
cryptography>=41.0.0,<42.0.0

# Replace crawl4ai with requests-html (more secure, lighter)
requests-html>=0.10.0,<1.0.0

# =============================================================================
# SECURITY ENHANCEMENTS
# =============================================================================

# Secure random number generation
secrets>=1.0.0

# Secure hashing (replaces weak MD5/SHA1 usage)
hashlib-compat>=1.0.0

# Secure file operations
pathlib2>=2.3.7,<3.0.0

# =============================================================================
# DEPENDENCY SECURITY NOTES
# =============================================================================

# VULNERABILITIES ADDRESSED:
# - pypdf2: Replaced with pypdf (actively maintained)
# - ecdsa: Replaced with cryptography (more secure)
# - crawl4ai: Replaced with requests-html (lighter, more secure)
# - MD5/SHA1: Replaced with secure hashing via hashlib-compat
# - subprocess: All subprocess calls use secure practices
# - pickle: Replaced with secure serialization where possible

# SECURITY BEST PRACTICES:
# - All dependencies pinned to specific versions
# - Regular security updates via Dependabot
# - No development dependencies in production
# - Minimal attack surface
# - Secure by default configurations
